refactor(intelligence,http): move ai_handler to mockforge-intelligence (#656)
Follow-up to #555 — `ai_handler` was the second-largest AI-touching
module still in mockforge-http (~274L). Its imports were
`mockforge_core::{Result, Error}` re-exports (already resolving to
`mockforge-foundation`) plus `mockforge_data` types (already a direct
dep of mockforge-intelligence), so the move stays cycle-safe with the
Issue #562 core↔intelligence cycle-break.

Changes:
- `mockforge-intelligence/src/ai_handler.rs` (new) — verbatim move
  with imports rewritten to `mockforge_foundation::{Error, Result}`
- `mockforge-http/src/ai_handler.rs` — now a 20-line re-export shim
  so `mockforge_http::ai_handler::{AiResponseHandler, AiResponseConfig,
  process_response_with_ai, create_ai_handler}` callers (notably the
  top-level `mockforge_http::{AiResponseConfig, AiResponseHandler,
  process_response_with_ai}` re-exports in lib.rs) keep resolving
  unchanged
- `mockforge-http/src/lib.rs` — updated the post-#555 commentary block
  to add `ai_handler` to the moved-out list

Net: -255 LOC from mockforge-http, +275 LOC to mockforge-intelligence.

Verified:
- cargo build -p mockforge-intelligence -p mockforge-http
- cargo clippy -p mockforge-intelligence -p mockforge-http --all-targets -- -D warnings
- cargo test -p mockforge-intelligence (343/343 pass)
- cargo test -p mockforge-http (571/571 pass)
- cargo fmt --all --check

Refs #656. `management/ai_gen.rs` (~994L, the other file scoped to
#656) is the larger remaining candidate; it needs a split decision
because it's coupled to `ManagementState`. Left for a follow-up PR.
